+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 0.6.9
+
+* added the following document types that can be created with Document.add(): 
+ 'Analysis', 'CatalogDocument', 'CATMaterial', 'CATProcess', 'cgm' 
+ 'FeatureDictionary', 'gl', 'gl2', 'hpgl', 'FunctionalSystem', 'ProcessLibrary'
+* Rewrote how document types are detected and tested for. This makes adding 
+ additional document types easier.
+* fixed HybridShapeExtrapol.support. @HyberCa
+* fixed HybridShapeAxisLine.element.
+
+
 ## 0.6.8
 
 * added option to initialise the Application object with pythoncom.CoInitialize().